Fix edge cases where in-app push notification doesn't show in-app
If the device is on low power, or background fetch is disabled, the push
notification wasn't loading the in-app message content prior to presentation.

This meant the message was sometimes not shown to the user.

Now we check if we could find the message locally, and if not, we will trigger
a sync to load the in-app message content.