Fix background event tracking race condition
Sometimes the background event would fail to flush the event queue, meaning it would be sent to
the server on the next foreground.

This happens because the 'asynchronously' flag on the event tracking method did not properly make
execution synchronous and blocking, so the background task was being invalidated before the network
requests were completed, suspending the thread handling transmission.

To fix, the asynchronous flag has been replaced with a callback which allows the task invalidation
to happen once the sync has been completed.

Because the session helper was the only consumer of the asynchronous flag, this is assumed to be a
low impact change.