Refactor push token request flow to account for application state and existing permissions
It has been observed on iOS 13.4 that push prompts can be shown (1) whilst the app is not
in the foreground, and (2), multiple times ot the same user.

For (1), we guard asking permission with application status being active.

For (2), we attempt to determine existing authorization status prior to requesting authorization
again. It is unclear if this will help as requestAuthorization is already meant to only ask once
according to the Apple Developer Guide, so this pre-flight check seems unnecessary.

Relevant sections:

- https://developer.apple.com/documentation/usernotifications/asking_permission_to_use_notifications?language=objc
- https://developer.apple.com/documentation/usernotifications/unusernotificationcenter/1649527-requestauthorizationwithoptions?language=objc
- https://developer.apple.com/documentation/usernotifications/unusernotificationcenter/1649524-getnotificationsettingswithcompl?language=objc