Handle error cases consistently from fingerprinting match failures
Only invoke the deep link handler with a failure for:

- Expiry
- Limit exceeded

Other cases can't be handled due to inability to correctly know the URL that
wasn't matched.

In addition, if the success handler fails to parse the expected URL from the
response, quietly abort too. API should always pass this field.