Support base paths in base URLs
The implementation of this has been updated to respect query string parameters which are currently passed in as part of the path from a few callsites. The previous solution using appendingPathComponent was encoding the whole path again, and this was messing up the query section.

There are limited call sites inthe SDK that pass query string parameters, in future these could be refactored to afford an API that allows passing a map of parameters and building of URLs using URLComponents.